requirements.txt
google-trans-new==1.1.9
requests
discord.py
python-dotenv
beautifulsoup4
pytz==2021.1